refactor: Домашнее задание #3 поправил компонент ошибки и отключил проверку типа пропса у Main
