+++ src/containers/App.tsx
@@ -1,20 +1,21 @@
-import { ErrorBoundary } from '@components/ErrorBoundary'
 import { Footer } from '@components/Footer'
 import { Header } from '@components/Header/Header'
 import { Main } from '@components/Main/Main'
+import { ErrorBoundary } from '@containers/ErrorBoundary'
 import * as React from 'react'
 import { getCards } from '~/data/data'
 
-export const App = (): JSX.Element => {
-	const cards = getCards()
+const cards = getCards()
 
+const AppComponent = (): JSX.Element | null => {
 	return (
 		<>
-			<ErrorBoundary data={cards}>
-				<Header />
-				<Main data={cards} />
-				<Footer />
-			</ErrorBoundary>
+			<Header />
+			{/*@ts-ignore*/}
+			<Main data={cards} />
+			<Footer />
 		</>
 	)
 }
+
+export const App = ErrorBoundary(AppComponent, cards)
